Rename DjangoProcess to KolibriHttpProcess
This name is clearer, particularly in light of only using one process.

https://phabricator.endlessm.com/T32972
